+++ app/src/main/java/com/eziby/eziby_android_app/Adapters/WishListAdapter.java
@@ -10,6 +10,7 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.eziby.eziby_android_app.Classes.EziByValues;
@@ -52,22 +53,27 @@ public class WishListAdapter extends RecyclerView.Adapter<WishListAdapter.ImageV
         Picasso.get()
                 .load(imageUri) // Image URL
                 .placeholder(R.drawable.loading_image_light_grey_100) // Placeholder image while loading
-                .error(R.drawable.error_image_30) // Error image if the URL fails to load
+//                .error(R.drawable.error_image_30) // Error image if the URL fails to load
                 .into(holder.product_image); // Target ImageView
 
         holder.product_name.setText(Optional.ofNullable(itemArray.get(position).getItemName()).orElse(""));
         holder.product_price.setText(_sellingPrice);
 
-        holder.delete_button.setOnClickListener(v -> {
-            try (DbHelper dbHelper = new DbHelper(context)) {
-                dbHelper.deleteWishList(itemArray.get(position).getWishListId());
-                // Remove the item from the list
-                itemArray.remove(position);
-                // Notify the adapter about the removed item
-                notifyItemRemoved(position);
-                notifyItemRangeChanged(position, itemArray.size());
-            }
-        });
+        holder.delete_button.setOnClickListener(v -> new AlertDialog.Builder(context)
+                .setTitle("Delete")
+                .setMessage("Remove Item?")
+                .setIcon(android.R.drawable.ic_delete)
+                .setPositiveButton(android.R.string.yes, (dialog, whichButton) -> {
+                    try (DbHelper dbHelper = new DbHelper(context)) {
+                        dbHelper.deleteWishList(itemArray.get(position).getWishListId());
+                        // Remove the item from the list
+                        itemArray.remove(position);
+                        // Notify the adapter about the removed item
+                        notifyItemRemoved(position);
+                        notifyItemRangeChanged(position, itemArray.size());
+                    }
+                })
+                .setNegativeButton(android.R.string.no, null).show());
 
         holder.add_to_shopping_cart_button.setOnClickListener(v -> {
             try (DbHelper dbHelper = new DbHelper(context)) {
